requirements.txt
PyMySQL==1.0.2
pytest==6.2.5
pytest-cov==3.0.0
pytest-cov==3.0.0
python-dotenv==0.19.1
pycodestyle==2.8.0